# src/main/resources/openapi/cyberwrite/apis.yml
components:
  schemas:
    APIErrorSchema:
      properties:
        errorMessage:
          title: Errormessage
          type: string
        request_id:
          title: Request Id
          type: string
      required:
      - request_id
      - errorMessage
      title: APIErrorSchema
      type: object
...
    AggregatedRiskInProgressSchema:
      properties:
        message:
          title: Message
          type: string
        success:
          default: true
          title: Success
          type: boolean
      required:
      - message
      title: AggregatedRiskInProgressSchema
      type: object
...
    InsuredsListSchema:
      properties:
        data:
          description: List of insureds
          items:
            $ref: '#/components/schemas/InsuredSchema'
          title: Data
          type: array
        pagination:
          allOf:
          - $ref: '#/components/schemas/PaginationSchema'
          description: Pagination information for the list of insureds
          title: Pagination
      required:
      - data
      - pagination
      title: InsuredsListSchema
      type: object
    InsuredsReprofileSchema:
      properties:
        action:
          default: reprofile
          title: Action
          type: string
        request_id:
          title: Request Id
          type: string
        success:
          default: true
          title: Success
          type: boolean
      required:
      - request_id
      title: InsuredsReprofileSchema
      type: object
...
info:
  description: "# Overview\n\nThe Cyberwrite API provides access to Cyberwrite\u2019\
    s Cyberprofile\u2122 technology which enables insurers, brokers, and agents to\
    \ manage the cyber insurance risk of small-medium size businesses and to estimate\
    \ the financial impact of a cyber incident on businesses.\n\n# Access Control\
    \ and Authorization\n\nAccess to the API is controlled by API keys issued to customers\
    \ per specific environment. So the key issued for a production environment will\
    \ not grant access to the test environment and vice versa.\n\nEach insurer will\
    \ be authorized to access the API for their resources only based on the insurer\
    \ element of the path. The authorization policy is evaluated based on the match\
    \ between API key, environment, and the insurer element of the path."
  title: Cyberwrite API Reference
  version: v2.0.8
openapi: 3.0.0
paths:
  /insurers/{insurer}/aggregated-risk:
    get:
      description: Get Aggregated Risk Report
      parameters:
      - description: The insurance company ID provided by Cyberwrite.
        example: '{{x-api-name}}'
        in: path
        name: insurer
        required: true
        schema:
          type: string
      - description: API key provided by Cyberwrite
        example: '{{x-api-key}}'
        in: header
        name: x-api-key
        required: true
        schema:
          type: string
      - default: null
        description: Book to filter Reports
        example: null
        in: query
        name: book
        required: false
        schema:
          type: string
      - default: null
        description: Agency ID to filter Reports
        example: null
        in: query
        name: agency
        required: false
        schema:
          type: string
      - default: '1970-01-01'
        description: Filter Reports since this date
        example: '2024-01-01'
        in: query
        name: start_date
        required: false
        schema:
          type: date
      - default: current date
        description: Filter Reports up to this date
        example: '2025-01-01'
        in: query
        name: end_date
        required: false
        schema:
          type: date
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregatedRiskSchema'
          description: Get Aggregated Risk Report
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        202:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregatedRiskInProgressSchema'
          description: The insurer aggregated risk data queued for processing
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        400:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The request does not adhere to rules defined by the API
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        401:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: Authorization error (wrong API key or user verification error)
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        403:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: Authorization error (wrong API key or user verification error)
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        404:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The requested resource does not exist
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        405:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The requested method not allowed
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        500:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: 'Error on the server-side (contact: support@cyberwrite.com)'
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        502:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: 'Unreachable backend service (contact: support@cyberwrite.com)'
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
      summary: Get Aggregated Risk Report
      tags:
      - Aggregated Risk API
  /insurers/{insurer}/insureds:
    get:
      description: Search Insureds by filter
      parameters:
      - description: The insurance company ID provided by Cyberwrite.
        example: '{{x-api-name}}'
        in: path
        name: insurer
        required: true
        schema:
          type: string
      - description: API key provided by Cyberwrite
        example: '{{x-api-key}}'
        in: header
        name: x-api-key
        required: true
        schema:
          type: string
      - default: 100
        description: Number of items per page
        example: 100
        in: query
        name: limit
        required: false
        schema:
          type: integer
      - default: 0
        description: Number of items to skip
        example: 0
        in: query
        name: skip
        required: false
        schema:
          type: integer
      - default: null
        description: Insured customer assigned ID
        example: null
        in: query
        name: assigned_id
        required: false
        schema:
          type: string
      - default: null
        description: String to search by website or name
        example: null
        in: query
        name: search
        required: false
        schema:
          type: string
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsuredsListSchema'
          description: Insured successfully received
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        400:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The request does not adhere to rules defined by the API
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        401:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: Authorization error (wrong API key or user verification error)
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        403:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: Authorization error (wrong API key or user verification error)
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        404:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The requested resource does not exist
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        405:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The requested method not allowed
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        500:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: 'Error on the server-side (contact: support@cyberwrite.com)'
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        502:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: 'Unreachable backend service (contact: support@cyberwrite.com)'
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
      summary: Search Insureds by filter
      tags:
      - Insured API
...
  /insurers/{insurer}/insureds/{insured}/reprofile:
    post:
      description: Reprofile Insured
      parameters:
      - description: The insurance company ID provided by Cyberwrite.
        example: '{{x-api-name}}'
        in: path
        name: insurer
        required: true
        schema:
          type: string
      - description: The insured UUID as defined in RFC4122
        example: '{{insured-id}}'
        in: path
        name: insured
        required: true
        schema:
          type: string
      - description: API key provided by Cyberwrite
        example: '{{x-api-key}}'
        in: header
        name: x-api-key
        required: true
        schema:
          type: string
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsuredsReprofileSchema'
          description: Reprofile Insured initiated
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        400:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The request does not adhere to rules defined by the API
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        401:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: Authorization error (wrong API key or user verification error)
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        403:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: Authorization error (wrong API key or user verification error)
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        404:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The requested resource does not exist
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        405:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The requested method not allowed
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        500:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: 'Error on the server-side (contact: support@cyberwrite.com)'
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        502:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: 'Unreachable backend service (contact: support@cyberwrite.com)'
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
      summary: Reprofile Insured
      tags:
      - Insured API
  /insurers/{insurer}/reports:
    get:
      description: Get Report Snapshots for Insurer
      parameters:
      - description: The insurance company ID provided by Cyberwrite.
        example: '{{x-api-name}}'
        in: path
        name: insurer
        required: true
        schema:
          type: string
      - description: API key provided by Cyberwrite
        example: '{{x-api-key}}'
        in: header
        name: x-api-key
        required: true
        schema:
          type: string
      - default: '1970-01-01T00:00:00'
        description: Get Snapshots since this date
        example: '2025-01-01T00:00:00'
        in: query
        name: since
        required: false
        schema:
          type: timestamp
      - default: 100
        description: Number of items per page
        example: 100
        in: query
        name: limit
        required: false
        schema:
          type: integer
      - default: 0
        description: Number of items to skip
        example: 0
        in: query
        name: skip
        required: false
        schema:
          type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsListSchema'
          description: Get Report Snapshots for Insurer
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        400:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The request does not adhere to rules defined by the API
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        401:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: Authorization error (wrong API key or user verification error)
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        403:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: Authorization error (wrong API key or user verification error)
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        404:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The requested resource does not exist
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        405:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The requested method not allowed
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        500:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: 'Error on the server-side (contact: support@cyberwrite.com)'
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        502:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: 'Unreachable backend service (contact: support@cyberwrite.com)'
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
      summary: Get Report Snapshots for Insurer
      tags:
      - Reports API
...
  /insurers/{insurer}/settings:
    get:
      description: Get Insurer or Agency settings
      parameters:
      - description: The insurance company ID provided by Cyberwrite.
        example: '{{x-api-name}}'
        in: path
        name: insurer
        required: true
        schema:
          type: string
      - description: API key provided by Cyberwrite
        example: '{{x-api-key}}'
        in: header
        name: x-api-key
        required: true
        schema:
          type: string
      - default: null
        description: Agency ID provided by Cyberwrite
        example: null
        in: query
        name: agency
        required: false
        schema:
          type: string
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsurerSettingsSchema'
          description: Insurer settings successfully received
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        400:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The request does not adhere to rules defined by the API
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        401:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: Authorization error (wrong API key or user verification error)
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        403:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: Authorization error (wrong API key or user verification error)
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        404:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The requested resource does not exist
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        405:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: The requested method not allowed
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        500:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: 'Error on the server-side (contact: support@cyberwrite.com)'
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
        502:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorSchema'
          description: 'Unreachable backend service (contact: support@cyberwrite.com)'
          headers:
            Content-Type:
              description: application/json
              schema:
                type: string
      summary: Get Insurer or Agency settings
      tags:
      - Insurer API
servers:
- description: Test server (uses test data)
  url: https://api-test.cyberwrite.com/v2.0.8
...